Add examples and some doc
